+++ demos/chinook/src/test/java/is/codion/demos/chinook/model/AlbumModelTest.java
@@ -58,7 +58,7 @@ public final class AlbumModelTest {
 			// to populate it with only Master Of Puppets
 			AlbumModel albumModel = new AlbumModel(connectionProvider);
 			SwingEntityTableModel albumTableModel = albumModel.tableModel();
-			albumTableModel.queryModel().conditions().setEqualOperand(Album.TITLE, MASTER_OF_PUPPETS);
+			albumTableModel.queryModel().conditions().get(Album.TITLE).set().equalTo(MASTER_OF_PUPPETS);
 			albumTableModel.items().refresh();
 			assertEquals(1, albumTableModel.items().count());