+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityTableModelTest.java
@@ -246,8 +246,7 @@ public abstract class AbstractEntityTableModelTest<EditModel extends EntityEditM
 		assertEquals(1, counter.get());
 		commissionModel.enabled().set(false);
 		assertEquals(2, counter.get());
-		commissionModel.operator().set(Operator.GREATER_THAN_OR_EQUAL);
-		commissionModel.operands().lower().set(1200d);
+		commissionModel.set().greaterThanOrEqualTo(1200d);
 		//automatically set enabled when upper bound is set
 		assertEquals(3, counter.get());
 		empModel.queryModel().conditionChanged().removeListener(conditionChangedListener);