+++ framework/model/src/main/java/is/codion/framework/model/EntityTableConditionModel.java
@@ -31,7 +31,6 @@ import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.framework.domain.entity.condition.Condition;
 
-import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Supplier;
@@ -51,26 +50,6 @@ public interface EntityTableConditionModel {
 	 * @return the connection provider
 	 */
 	EntityConnectionProvider connectionProvider();
-
-	/**
-	 * Sets the EQUAL condition operand of the condition model associated with {@code attribute}.
-	 * Enables the condition model in case {@code operand} is non-empty or disables it if {@code operand} is empty.
-	 * @param attribute the attribute
-	 * @param operand the search condition operand
-	 * @param <T> the operand type
-	 * @return true if the search state changed as a result of this method call, false otherwise
-	 */
-	<T> boolean setEqualOperand(Attribute<T> attribute, T operand);
-
-	/**
-	 * Sets the IN condition operands of the condition model associated with {@code attribute}.
-	 * Enables the condition model in case {@code operands} is non-empty or disables it if {@code operands} is empty.
-	 * @param attribute the attribute
-	 * @param operands the search condition operands, an empty Collection for none
-	 * @param <T> the value type
-	 * @return true if the search state changed as a result of this method call, false otherwise
-	 */
-	<T> boolean setInOperands(Attribute<T> attribute, Collection<T> operands);
 
 	/**
 	 * Returns a WHERE condition based on enabled condition models which are based on non-aggregate function columns.